Safari-609.1.11.1/LayoutTests/platform/ios/media/mediacapabilities/mediacapabilities-encodingInfo-undefined.html

<!DOCTYPE html>
<html>
<head>
	<title>mediacapabilities-encodingInfo-empty</title>
	<script src="../../../../resources/js-test-pre.js"></script>
	<script>

	</script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Test that MediaCapabilities.prototype.encodingInfo is missing.');
shouldBe('typeof navigator.mediaCapabilities.encodingInfo', '"undefined"');
shouldBe('typeof MediaCapabilities.prototype.encodingInfo', '"undefined"');
</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
