LayoutTests/platform/mac/media/mediacapabilities/mediacapabilities-encodingInfo-undefined-expected.txt

Test that MediaCapabilities.prototype.encodingInfo is missing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof navigator.mediaCapabilities.encodingInfo is "undefined"
PASS typeof MediaCapabilities.prototype.encodingInfo is "undefined"
PASS successfullyParsed is true

TEST COMPLETE
